feat: add support for selecting tilesheet
This commit adds support for selecting a tilesheet to be used in the
game. It includes changes to the Redux store, actions, and reducer to
manage the selected tilesheet. It also includes changes to the MenuScene
to display the selected tilesheet and toggle the menu sidebar. Finally,
it includes changes to the Game component to use the selected tilesheet
in the GridScene.